+++ app/Http/Controllers/Bible/GetBibleController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Bible;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Book;
+use App\Models\Chapter;
 use App\Models\Verse;
 
 class GetBibleController extends Controller
@@ -15,8 +16,25 @@ class GetBibleController extends Controller
     }
 
     public function getBibleChapter($book_id, $chapter) {
+        //TODO: This query needs to change to look up the chapter_id from the chapters table.
+
+        $chapter_row = Chapter::where('book_id', $book_id)
+                        ->where('book_chapter', $chapter)
+                        ->get();
+
+        $chapter_id = null;
+        foreach($chapter_row as $chap) {
+            $chapter_id = $chap->id;
+        }
+
+
         return Verse::where('book_id', $book_id)
-            ->where('chapter', $chapter)
+            ->where('chapter_id', $chapter_id)
+            ->get();
+    }
+
+    public function getChapterId($chapter_id) {
+        return Chapter::where('id', $chapter_id)
             ->get();
     }