+++ resources/js/components/Bible/LeafComponent.vue
@@ -2,11 +2,23 @@
     <div class="container">
         <div class="row justify-content-center">
             <div class="col-md-8">
-                <div class="card">
+                <div class="card" v-on:click="clickLeaf()">
                     <div class="card-header">Leaf Component</div>
 
                     <div class="card-body">
-                        I'm a {{ type }} Leaf Component.
+                        <span v-if="type==-1">
+                            <svg width="3em" height="2em" viewBox="0 0 16 16" class="bi bi-arrow-left" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
+                                <path fill-rule="evenodd" d="M15 8a.5.5 0 0 0-.5-.5H2.707l3.147-3.146a.5.5 0 1 0-.708-.708l-4 4a.5.5 0 0 0 0 .708l4 4a.5.5 0 0 0 .708-.708L2.707 8.5H14.5A.5.5 0 0 0 15 8z"/>
+                            </svg>
+                        </span>
+                        
+                        <span v-if="type==1">
+                            <svg width="3em" height="2em" viewBox="0 0 16 16" class="bi bi-arrow-right" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
+                                <path fill-rule="evenodd" d="M1 8a.5.5 0 0 1 .5-.5h11.793l-3.147-3.146a.5.5 0 0 1 .708-.708l4 4a.5.5 0 0 1 0 .708l-4 4a.5.5 0 0 1-.708-.708L13.293 8.5H1.5A.5.5 0 0 1 1 8z"/>
+                            </svg>
+                        </span>
+                        
+
                     </div>
                 </div>
             </div>
@@ -16,8 +28,10 @@
 
 <script>
     export default {
-        mounted() {
-            console.log('Component mounted.')
+        methods: {
+            clickLeaf: function() {
+                this.$emit('leaf-page', this.type);
+            }
         },
         props:['type']
     }