+++ resources/js/components/Bible/PageComponent.vue
@@ -8,11 +8,20 @@
             <p><span class="h5">{{ index + 1 }}: </span><span v-html="text.verse"></span></p>
         </div>
 
+        <leaf-component type=-1 v-on:leaf-page="leafPage"></leaf-component>
+        <leaf-component type=1 v-on:leaf-page="leafPage"></leaf-component>
+
+
     </div>
 </template>
 
 <script>
     export default {
+        methods: {
+            leafPage: function (_type){
+                this.$emit('leaf-page', _type);
+            }
+        },
         props: ['page', 'chapterText']
     }
 </script>